[IOTDB-2122] Migrate MQTT and REST modules to `org.apache.iotdb.db.protocol` (#4564)
