+++ example/mqtt-customize/src/main/resources/META-INF/services/org.apache.iotdb.db.org.apache.iotdb.db.protocol.mqtt.PayloadFormatter
