+++ server/src/main/java/org/apache/iotdb/db/protocol/rest/handler/RequestValidationHandler.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.iotdb.db.protocol.rest.handler;
+
+import org.apache.iotdb.db.protocol.rest.model.InsertTabletRequest;
+import org.apache.iotdb.db.protocol.rest.model.SQL;
+
+import java.util.Objects;
+
+public class RequestValidationHandler {
+  private RequestValidationHandler() {}
+
+  public static void validateSQL(SQL sql) {
+    Objects.requireNonNull(sql.getSql(), "sql should not be null");
+  }
+
+  public static void validateInsertTabletRequest(InsertTabletRequest insertTabletRequest) {
+    Objects.requireNonNull(insertTabletRequest.getTimestamps(), "timestamps should not be null");
+    Objects.requireNonNull(insertTabletRequest.getIsAligned(), "isAligned should not be null");
+    Objects.requireNonNull(insertTabletRequest.getDeviceId(), "deviceId should not be null");
+    Objects.requireNonNull(insertTabletRequest.getDataTypes(), "dataTypes should not be null");
+    Objects.requireNonNull(insertTabletRequest.getValues(), "values should not be null");
+  }
+}
